Rename omeroweb/http.py to avoid import Errors
